Extract chat completions conversion code into helper
