+++ tests/test_tool_converter.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 from agents import Agent, Handoff, function_tool, handoff
 from agents.exceptions import UserError
-from agents.models.openai_chatcompletions import ToolConverter
+from agents.models.chatcmpl_converter import Converter
 from agents.tool import FileSearchTool, WebSearchTool
 
 
@@ -15,7 +15,7 @@ def test_to_openai_with_function_tool():
     some_function(a="foo", b=[1, 2, 3])
 
     tool = function_tool(some_function)
-    result = ToolConverter.to_openai(tool)
+    result = Converter.tool_to_openai(tool)
 
     assert result["type"] == "function"
     assert result["function"]["name"] == "some_function"
@@ -34,7 +34,7 @@ class Foo(BaseModel):
 def test_convert_handoff_tool():
     agent = Agent(name="test_1", handoff_description="test_2")
     handoff_obj = handoff(agent=agent)
-    result = ToolConverter.convert_handoff_tool(handoff_obj)
+    result = Converter.convert_handoff_tool(handoff_obj)
 
     assert result["type"] == "function"
     assert result["function"]["name"] == Handoff.default_tool_name(agent)
@@ -48,7 +48,7 @@ def test_convert_handoff_tool():
 
 def test_tool_converter_hosted_tools_errors():
     with pytest.raises(UserError):
-        ToolConverter.to_openai(WebSearchTool())
+        Converter.tool_to_openai(WebSearchTool())
 
     with pytest.raises(UserError):
-        ToolConverter.to_openai(FileSearchTool(vector_store_ids=["abc"], max_num_results=1))
+        Converter.tool_to_openai(FileSearchTool(vector_store_ids=["abc"], max_num_results=1))
